requirements.txt
# Core Web Framework
Django>=5.2
djangorestframework>=3.16
django-filter>=23.0
django-cors-headers>=4.0.0
django-rest-framework-mongoengine>=3.4.1
djangorestframework-simplejwt>=5.5.1


# Documentação e Deploy
drf-yasg>=1.21.11
gunicorn>=21.0.0
whitenoise>=6.0.0


# Banco de Dados e Armazenamento
mongoengine>=0.29.1
pymongo>=4.15.3
boto3>=1.40.46


# ETL e Processamento de Dados
pandas>=2.3.3
fastparquet>=2024.11.0
pyspark>=4.0.1


# Utilitários e Integrações
requests>=2.32.5
beautifulsoup4>=4.14.2
python-decouple>=3.8
dnspython>=2.8.0
black>=25.9.0


# Jupyter
ipykernel>=6.30.1
jupyter>=1.1.1
